test: split large files
